Updated the repository structure to comply with SE 510 Homework's layout Renamed src/ to code/ Renamed tests/ to test/ Added test/__init__.py Modified setup.py after changing src/ to code/
